tests/ztest/fail: Fix build for native_sim
The final executable output from the native_sim
(or native_posix) build is zephyr.exe
(in native_posix zephyr.elf happened to be just a copy of
the exe, but in native_sim it is an intermediate build step)

So we need to use the exe instead.

to install it using install(FILES), which also requires the
destination and permissions to be listed manually. The
permissions set are just the cmake default permissions
for install(PROGRAMS).

(cherry picked from commit 0b3a7bc7d4b7c1e242956b67449fca933444a4e4)

Original-Note: As the exe is generated with a custom target, we need
Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 0b3a7bc7d4b7c1e242956b67449fca933444a4e4
Change-Id: I5f693827f66258427c13a6ce91ca7660b35f2175
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5049280
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>