drivers: clock_control: clock_control_ra.c: clock divider fix
The clock divider value is not being applied as the address of the
register to which it is being written is incorrect. A check of all RA MCU
datasheets confirms that, in all cases, the SCKDIVCR register is at an
offset of 0x20 (and not 0x21).

(cherry picked from commit 640e70c9ce5f8bd5b2028533d38f52f8a8706c73)

Original-Signed-off-by: Ian Morris <ian.d.morris@outlook.com>
GitOrigin-RevId: 640e70c9ce5f8bd5b2028533d38f52f8a8706c73
Change-Id: Ibb750c2d64a5454526e56c3a10c4e86cd4b1cfb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5216997
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>