bmi160: fix value calculation
As an example, when the gyro range is set to 250 deg/sec the scale
was set to 133 where it really should be 133.160058662. This leads
to a 0.12% error in the value returned. By separating the numerator
and denominator, we're able to drastically reduce the error.

(cherry picked from commit 9381e86cbe0d7dd00de4c74d060f7b8d7661c7ee)

Original-Signed-off-by: Yuval Peress <peress@google.com>
GitOrigin-RevId: 9381e86cbe0d7dd00de4c74d060f7b8d7661c7ee
Change-Id: Ibb757f0de0cfd5bec544320eaf57044dcec0084c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5200457
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>