drivers: i3c: cdns: handle controller aborts
Some targets do not give EoD at the end of a register read. They will
auto increment their address pointer on to the next address, but that
may not be of interest to the application where the buffer size will
only be set to the size of only that register. If the target, does
not give an EoD, then the Controller will give an Abort... but this
should not be treated as an error in this case.

There is still however a case where an abort Error shall still be
considered as an error. Athough the driver does not support it yet,
threshold interrupts are to be used if the length of the buffer
exceeds the size of the fifo. There could be the case where the
cpu can not get around fast enough to pop out data out of the rx
fifo and it will fill up. The controller will just give an abort
as it can not take any more data.

(cherry picked from commit 9c22c82604763021004606d2e54a3ee81715c9b2)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: 9c22c82604763021004606d2e54a3ee81715c9b2
Change-Id: Ibc42a5c4af58a83cb6e1cadff1fb03b087045b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5129349
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>