soc: gd32vf103: Correct vector table alignment
For a CLIC the vector table has to be aligned by 512 bytes, if there are
between 65 and 128 interrupts, which is the case for the gd32vf103.

`isr_wrapper` has to be aligned to 64 bytes, s.t. the lower 6 bits of
mtvec are 0.

(cherry picked from commit 87916d7af39616705192f8a5d2f01177f316f35f)

Original-Signed-off-by: Greter Raffael <rgreter@baumer.com>
GitOrigin-RevId: 87916d7af39616705192f8a5d2f01177f316f35f
Change-Id: Id71e208186367d51c82ecf8b34553e238f1bd46a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5200444
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>