tfm: Update prompt for profile not set
Update prompt for profile type not set, make it appear in a similar
way as the rest of the options in the choice.

Add in parethesis "base", which is how this referred to in the upstream
TF-M documentation.

(cherry picked from commit 44106dac56b530d21f3682bfe1f215349f64e3de)

Original-Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
GitOrigin-RevId: 44106dac56b530d21f3682bfe1f215349f64e3de
Change-Id: I992348459c8f173eb60eb20e4d9e7da42c08aa49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5115755
Tested-by: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>