drivers: intc_plic: enable contexts for all harts on a platform
The plic uses contexts to seperate irq enables, threshold priority and
claim complete registers from each core for a given platform. As well as
this, each privilege level has its own context.
for multi-core platform's, we need to be able to enable/ disable a
global interrupt for all the cores that are associated with Zephyr.

To do this, we need to make some assumptions:
1. The privilege contexts are contiguous
2. M mode context is first, followed by S mode.

We know how many cpus are used in an application and each cpu's hartid,
thanks to some very handy inline functions. So we iterate through each
cpu and use the hartid of a cpu in the calculation of the context.

While we are at it, In an effort to make the driver more readable,
allign with the macro naming convention outlined in Linux's PLIC driver

(cherry picked from commit 4e825f572c004c71754c0cd6c906afdfa0e98ed8)

Original-Signed-off-by: Conor Paxton <conor.paxton@microchip.com>
GitOrigin-RevId: 4e825f572c004c71754c0cd6c906afdfa0e98ed8
Change-Id: Ic312bb81d85d9659898f1893c3396507e921c033
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5115795
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>