drivers: intc_plic: claim interrupt from hart that serviced it
The plic has a very simple mechanism to claim an interrupt as well as to
complete and clear it. The same register is read from/ written to to
achieve this.

Get the ID of the HART that serviced the interrupt and write to the
claim complete register in the correct context
(cherry picked from commit 7e192e9e279cf3a0d25eed6d874ef7f9d3d4278d)

Original-Signed-off-by: Conor Paxton <conor.paxton@microchip.com>
GitOrigin-RevId: 7e192e9e279cf3a0d25eed6d874ef7f9d3d4278d
Change-Id: I8e896af32c889fe6e7a72d84aa3b35aa31c75a13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5115796
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>