devicetree: remove label property accessors
These were first deprecated in v3.2. We have kept support for them
for as long as we can, but now their presence is causing CI failures
in some configurations. Specifically, using the deprecated 'label'
property is causing a warning which is promoted to an error in some
twister runs. This is blocking other forward progress in the
devicetree API.

I tried to rework the tests to avoid this, but it was too much effort
for the time I had to work on the task. Removing the APIs is therefore
unfortunately the best way forward to unblocking other work.

Re-work the test suite a bit to maintain coverage where we are using
the label property to test other macros.

Add a migration guide section to help with the transition.

(cherry picked from commit b1532ce44914a9c7495d0bd1cdf198ed0e6d1cb8)

Original-Signed-off-by: Martí Bolívar <mbolivar@amperecomputing.com>
GitOrigin-RevId: b1532ce44914a9c7495d0bd1cdf198ed0e6d1cb8
Change-Id: If65e01e059a78c3e530a2bfe16eebe45a6b15114
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5081675
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>