drivers: sensor: mcp9600: fix wrong register address
The address of most of the registers defined in
the driver are wrong. This fixes it, following
the correct numbering as can be found in the
device's datasheet. Moreover, re-grouping of
the macros according to their functionality.

(cherry picked from commit ed610400bcb74338d25e4b29f948709efe12f3a3)

Original-Signed-off-by: Nikos Agianniotis <na@neq.dk>
GitOrigin-RevId: ed610400bcb74338d25e4b29f948709efe12f3a3
Change-Id: Ifac6ef60d53cfdb390db8e8207341e52e9777573
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5099397
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>