Bluetooth: ascs: Avoid possible unexpected assert
This avoids unexpected assert that may happen when the client tries to
QoS configure ASE that is in state which does not allow to be configured.
In such case the assert shall not be not be triggered, as it's not stack
fauly. The assert check has been moved after the state check, so the ASCS
implementation will just return an error code to the client.

(cherry picked from commit 025ba06c32976dc87a27680266d29a2c3cedcd2b)

Original-Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>
GitOrigin-RevId: 025ba06c32976dc87a27680266d29a2c3cedcd2b
Change-Id: I8c1bbdc3cdaa1dae6dd5cc4af8e1db99ec1b97da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5119809
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>