samples: soc_flash_nrf: Print finished message
Print a finished message, so that it is clear that this in the end
of the sample. This makes it certain that we didn't crash without
output on the last test.

(cherry picked from commit 29fc537bfbd0c4dc354d2c858e0e46d64393846b)

Original-Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
GitOrigin-RevId: 29fc537bfbd0c4dc354d2c858e0e46d64393846b
Change-Id: Ia2fa1f60603d93457e5cfc89f9abd68e8489d592
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5109837
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>