Revert "boards: arm: Fix sensor shell sample for beagleconnect_freedom board"
This reverts commit 3769938e785b09cff9ebb604a371100d1bfcf257.

Now that gpio_i2c_switch is upstream, this hack should be removed. This
had the unintended effect of disabling the on-board sensors. The issue
was that the board was merged before this driver was upstream, resulting
in the issue this commit "fixed".

This revert also does not move the bus driver under the controller used
by the driver as the sensor bus I2C driver is not an I2C peripheral.

See:
* https://github.com/zephyrproject-rtos/zephyr/pull/64881
* https://github.com/zephyrproject-rtos/zephyr/pull/64693
* https://github.com/zephyrproject-rtos/zephyr/pull/65031

(cherry picked from commit 547a75d4d05b09a925179d3f4fa477facde609bd)

Original-Signed-off-by: Jason Kridner <jkridner@beagleboard.org>
GitOrigin-RevId: 547a75d4d05b09a925179d3f4fa477facde609bd
Change-Id: Ib6a6a0f71d249bc19a312d469f6e41224b70524b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5119865
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>