pm: device_runtime: Avoid unnecessary work
Similar to what is done in pm_device_runtime_get,
we don't need to wait the device be suspended if
the work is still in the work queue. In this case
we just cancel the work and update the device state
to active.

(cherry picked from commit b16110587c76cebd644c01f9e45218412281e140)

Original-Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: b16110587c76cebd644c01f9e45218412281e140
Change-Id: I466c7f54847a454dc79f22cc19a36fb7565091fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5126131
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>