modules: hal_nordic: reorganize nrf_802154 Kconfig
There are multiple Kconfig symbols that are only available when the
radio hardware is present. As a result, the nRF 802.15.4 radio driver
is more difficult to configure on multicore architectures. Also, such
solution goes against the principle of hiding the details of the
platform behind the driver's interface.

This commit moves all the Kconfig options that correspond to settings
present in common nRF 802.15.4 configuration header so that they're
available no matter if the API is serialized.

(cherry picked from commit 054353a75daac5522a765251201e93b4a722882f)

Original-Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>
GitOrigin-RevId: 054353a75daac5522a765251201e93b4a722882f
Change-Id: I79349e6f8e88ba1c188f65a4d90b5bac82349800
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5163000
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>