bindesc: Fix placing of the descriptors end tag
Prior to this fix the descriptors end tag was placed in the binary
using LONG. This was wrong because the tag is of type short, and
the size of the descriptor (0), should have also been placed.
It so happened that on little endian machines the incorrect linker
script yielded correct results, as the extra zeros added to make
the value a long ended up as the correct byte sequence
(FF FF 00 00). On big endian machines however, the wrong sequence
is generated (00 00 FF FF).
This patch correct this issue.

(cherry picked from commit c8c25b5c52a9dee347eae34137559bdbe48e4505)

Original-Signed-off-by: Yonatan Schachter <yonatan.schachter@gmail.com>
GitOrigin-RevId: c8c25b5c52a9dee347eae34137559bdbe48e4505
Cr-Build-Id: 8734199375406094193
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8734199375406094193
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4dae79edf97cbad2eed34780722c10caaacf3c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5927192
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>