drivers: dai: sai: Introduce the rx_sync_mode/tx_sync_mode properties
In preparation for supporting all synchronization modes, this
commit introduces the rx_sync_mode/tx_sync_mode DTS properties.
Using these, the user will be able to specify which synchronization
mode the SAI should use.

At the moment, the driver does nothing with the values from
said properties but still checks if their values are sane
(i.e: it checks if the directions are both in SYNC mode which
is forbidden). By default, if "rx_sync_mode" or "tx_sync_mode"
is not specified, the direction will be set to ASYNC mode.
As such, below one may find a couple of valid examples
depicting this idea:

	tx_sync_mode = <0>;
	rx_sync_mode = <0>;

is the same as not specifying any of the properties,

	tx_sync_mode = <1>;
	rx_sync_mode = <0>;

is the same as:

	tx_sync_mode = <1>;

(cherry picked from commit 6127535b1d5a08df2ad4ab2df9c14fb8304fe226)

Original-Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
GitOrigin-RevId: 6127535b1d5a08df2ad4ab2df9c14fb8304fe226
Change-Id: I4b1b4e34e0f05cc324662d611dcc320f62f8fa99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5188963
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>